gui wip

+++ cleanup.js
@@ -25,7 +25,7 @@ exports.cleanup = (userArgs, path, fs, utils, dataDir)=>{
 				for (const file of files) {
 					let ext = path.extname(file);
 
-					if (!fs.lstatSync(path.join(dir, file)).isDirectory() && (ext !== ".hdb")) {
+					if (!fs.lstatSync(path.join(dir, file)).isDirectory() && (ext !== ".hdb") && (ext !== ".sdb")) {
 						(userArgs.CLEAN_ON_STARTUP && userArgs.DEBUG) && utils.console("Cleaning " + file);
 						fs.unlink(path.join(dir, file), err => {
 							//if (err) throw err;
